+++ openHAB/AppDelegate.swift
@@ -9,6 +9,7 @@
 //
 // SPDX-License-Identifier: EPL-2.0
 
+import AlamofireNetworkActivityIndicator
 import AVFoundation
 import Firebase
 import Kingfisher
@@ -55,7 +56,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         Preferences.migrateUserDefaultsIfRequired()
 
-        NetworkConnection.initialize(ignoreSSL: Preferences.ignoreSSL, adapter: OpenHABAccessTokenAdapter(appData: AppDelegate.appDelegate.appData))
+        NetworkConnection.initialize(ignoreSSL: Preferences.ignoreSSL, interceptor: OpenHABAccessTokenAdapter(appData: AppDelegate.appDelegate.appData))
+
+        NetworkActivityIndicatorManager.shared.isEnabled = true
+        NetworkActivityIndicatorManager.shared.startDelay = 1.0
 
         registerForPushNotifications()