+++ openHAB/OpenHABSettingsViewController.swift
@@ -190,8 +190,7 @@ class OpenHABSettingsViewController: UITableViewController, UITextFieldDelegate 
     }
 
     override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        let version = appData?.openHABVersion ?? 2
-        return version >= 3 && indexPath == IndexPath(row: 5, section: 1) ? .zero : UITableView.automaticDimension
+        UITableView.automaticDimension
     }
 
     func enableConnectionSettings() {