+++ openHAB/OpenHABViewController.swift
@@ -173,6 +173,7 @@ class OpenHABViewController: UIViewController {
                 UIAlertAction(title: NSLocalizedString("activate", comment: ""), style: .default) { _ in
                     Preferences.sendCrashReports = true
                     Crashlytics.crashlytics().setCrashlyticsCollectionEnabled(true)
+                    Crashlytics.crashlytics().sendUnsentReports()
                 }
             )
             alertController.addAction(
@@ -182,7 +183,11 @@ class OpenHABViewController: UIViewController {
                     self?.present(webViewController, animated: true)
                 }
             )
-            alertController.addAction(UIAlertAction(title: NSLocalizedString("cancel", comment: ""), style: .default))
+            alertController.addAction(
+                UIAlertAction(title: NSLocalizedString("cancel", comment: ""), style: .default) { _ in
+                    Crashlytics.crashlytics().deleteUnsentReports()
+                }
+            )
             present(alertController, animated: true)
         }
 
@@ -395,7 +400,7 @@ class OpenHABViewController: UIViewController {
     }
 
     func relevantWidget(indexPath: IndexPath) -> OpenHABWidget? {
-        relevantPage?.widgets[indexPath.row]
+        relevantPage?.widgets[safe: indexPath.row]
     }
 
     private func updateWidgetTableView() {
@@ -445,7 +450,7 @@ class OpenHABViewController: UIViewController {
             guard let self = self else { return }
 
             switch response.result {
-            case .success:
+            case let .success(data):
                 os_log("Page loaded with success", log: OSLog.remoteAccess, type: .info)
                 let headers = response.response?.allHeaderFields
 
@@ -454,34 +459,34 @@ class OpenHABViewController: UIViewController {
                     os_log("Found X-Atmosphere-tracking-id: %{PUBLIC}@", log: .remoteAccess, type: .info, NetworkConnection.atmosphereTrackingId)
                 }
                 var openHABSitemapPage: OpenHABSitemapPage?
-                if let data = response.result.value {
-                    // If we are talking to openHAB 1.X, talk XML
-                    if self.appData?.openHABVersion == 1 {
-                        let str = String(decoding: data, as: UTF8.self)
-                        os_log("%{PUBLIC}@", log: .remoteAccess, type: .info, str)
 
-                        guard let doc = try? XMLDocument(data: data) else { return }
-                        if let rootElement = doc.root, let name = rootElement.tag {
-                            os_log("XML sitemap with root element: %{PUBLIC}@", log: .remoteAccess, type: .info, name)
-                            if name == "page" {
-                                openHABSitemapPage = OpenHABSitemapPage(xml: rootElement)
-                            }
-                        }
-                    } else {
-                        // Newer versions talk JSON!
-                        os_log("openHAB 2", log: OSLog.remoteAccess, type: .info)
-                        do {
-                            // Self-executing closure
-                            // Inspired by https://www.swiftbysundell.com/posts/inline-types-and-functions-in-swift
-                            openHABSitemapPage = try {
-                                let sitemapPageCodingData = try data.decoded(as: OpenHABSitemapPage.CodingData.self)
-                                return sitemapPageCodingData.openHABSitemapPage
-                            }()
-                        } catch {
-                            os_log("Should not throw %{PUBLIC}@", log: OSLog.remoteAccess, type: .error, error.localizedDescription)
+                // If we are talking to openHAB 1.X, talk XML
+                if self.appData?.openHABVersion == 1 {
+                    let str = String(decoding: data, as: UTF8.self)
+                    os_log("%{PUBLIC}@", log: .remoteAccess, type: .info, str)
+
+                    guard let doc = try? XMLDocument(data: data) else { return }
+                    if let rootElement = doc.root, let name = rootElement.tag {
+                        os_log("XML sitemap with root element: %{PUBLIC}@", log: .remoteAccess, type: .info, name)
+                        if name == "page" {
+                            openHABSitemapPage = OpenHABSitemapPage(xml: rootElement)
                         }
                     }
+                } else {
+                    // Newer versions talk JSON!
+                    os_log("openHAB 2", log: OSLog.remoteAccess, type: .info)
+                    do {
+                        // Self-executing closure
+                        // Inspired by https://www.swiftbysundell.com/posts/inline-types-and-functions-in-swift
+                        openHABSitemapPage = try {
+                            let sitemapPageCodingData = try data.decoded(as: OpenHABSitemapPage.CodingData.self)
+                            return sitemapPageCodingData.openHABSitemapPage
+                        }()
+                    } catch {
+                        os_log("Should not throw %{PUBLIC}@", log: OSLog.remoteAccess, type: .error, error.localizedDescription)
+                    }
                 }
+
                 self.currentPage = openHABSitemapPage
                 if self.isFiltering {
                     self.filterContentForSearchText(self.search.searchBar.text)
@@ -501,7 +506,7 @@ class OpenHABViewController: UIViewController {
                 if (error as NSError?)?.code == -1001, longPolling {
                     os_log("Timeout, restarting requests", log: OSLog.remoteAccess, type: .error)
                     self.loadPage(false)
-                } else if (error as NSError?)?.code == -999 {
+                } else if error.isExplicitlyCancelledError {
                     os_log("Request was cancelled", log: OSLog.remoteAccess, type: .error)
                 } else {
                     // Error
@@ -549,8 +554,8 @@ class OpenHABViewController: UIViewController {
     func selectSitemap() {
         NetworkConnection.sitemaps(openHABRootUrl: openHABRootUrl) { response in
             switch response.result {
-            case .success:
-                self.sitemaps = deriveSitemaps(response.result.value, version: self.appData?.openHABVersion)
+            case let .success(data):
+                self.sitemaps = deriveSitemaps(data, version: self.appData?.openHABVersion)
                 switch self.sitemaps.count {
                 case 2...:
                     if !self.defaultSitemap.isEmpty {
@@ -629,8 +634,7 @@ class OpenHABViewController: UIViewController {
         openHABAlwaysSendCreds = Preferences.alwaysSendCreds
         defaultSitemap = Preferences.defaultSitemap
         idleOff = Preferences.idleOff
-        let rawIconType = Preferences.iconType
-        iconType = IconType(rawValue: rawIconType) ?? .png
+        iconType = IconType(rawValue: Preferences.iconType) ?? .png
 
         appData?.openHABUsername = openHABUsername
         appData?.openHABPassword = openHABPassword
@@ -658,14 +662,14 @@ class OpenHABViewController: UIViewController {
         if !pageUrl.isEmpty {
             let pageReachability = NetworkReachabilityManager(host: pageUrl)
             if !pageNetworkStatusAvailable {
-                pageNetworkStatus = pageReachability?.networkReachabilityStatus
+                pageNetworkStatus = pageReachability?.status
                 pageNetworkStatusAvailable = true
                 return false
             } else {
-                if pageNetworkStatus == pageReachability?.networkReachabilityStatus {
+                if pageNetworkStatus == pageReachability?.status {
                     return false
                 } else {
-                    pageNetworkStatus = pageReachability?.networkReachabilityStatus
+                    pageNetworkStatus = pageReachability?.status
                     return true
                 }
             }
@@ -727,31 +731,24 @@ extension OpenHABViewController: OpenHABTrackerDelegate {
 
         NetworkConnection.tracker(openHABRootUrl: openHABRootUrl) { response in
             switch response.result {
-            case .success:
+            case let .success(data):
+                do {
+                    self.serverProperties = try data.decoded(as: OpenHABServerProperties.self)
+                    os_log("This is an openHAB >= 2.X", log: .remoteAccess, type: .info)
+                    let version = Int(self.serverProperties?.version ?? "2") ?? 2
+                    self.appData?.openHABVersion = version
+                    self.selectSitemap()
 
-                if let data = response.result.value {
-                    do {
-                        self.serverProperties = try data.decoded(as: OpenHABServerProperties.self)
-                        os_log("This is an openHAB >= 2.X", log: .remoteAccess, type: .info)
-                        let version = Int(self.serverProperties?.version ?? "2") ?? 2
-                        self.appData?.openHABVersion = max(2, version)
-                        if version >= 3 {
-                            self.iconType = .svg
-                            Preferences.iconType = IconType.svg.rawValue
-                        }
+                } catch {
+                    os_log("Could not decode response as JSON, test for OH1", log: .notifications, type: .error, error.localizedDescription)
+                    let str = String(decoding: data, as: UTF8.self)
+                    if str.hasPrefix("<?xml") {
+                        self.appData?.openHABVersion = 1
                         self.selectSitemap()
-
-                    } catch {
-                        os_log("Could not decode response as JSON, test for OH1", log: .notifications, type: .error, error.localizedDescription)
-                        let str = String(decoding: data, as: UTF8.self)
-                        if str.hasPrefix("<?xml") {
-                            self.appData?.openHABVersion = 1
-                            self.selectSitemap()
-                        }
                     }
                 }
-
             case let .failure(error):
+                self.openHABTrackingError(error)
                 os_log("This is not an openHAB server", log: .remoteAccess, type: .info)
                 os_log("On Connecting %{PUBLIC}@ %d", log: .remoteAccess, type: .error, error.localizedDescription, response.response?.statusCode ?? 0)
             }
@@ -775,7 +772,7 @@ extension OpenHABViewController: OpenHABTrackerDelegate {
     }
 
     func openHABTrackingError(_ error: Error) {
-        os_log("OpenHABViewController discovery error", log: .viewCycle, type: .info)
+        os_log("Tracking error: %{PUBLIC}@", log: .viewCycle, type: .info, error.localizedDescription)
         var config = SwiftMessages.Config()
         config.duration = .seconds(seconds: 60)
         config.presentationStyle = .bottom