+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Delete unsent crashlytics reports if permission dialog is canceled
+- Re-enable icon type setting
+- Fix certificate issues
+- Fixed connection to myopenhab.org
+- OpenHABCore swift package
+- Migrate to Alamofire 5
+- Move external dependencies to SPM
+- Remove cocoapods
+
 ## [Version 2.4.15, Build 1580410464] - 2020-12-04
 
 ### Added